moved eraseAccidental logic to separate component
